requirements.txt
flask==3.0.0
google-cloud-speech==2.26.0
